refactor: add game fold
