+++ index.ts
@@ -48,7 +48,7 @@ wss.on('connection', (ws:CustomWebSocket)=>{
         break;
       case 'attack':
         userAttack(ws, receivedMessage);
-        turnUser(ws, receivedMessage);
+        // turnUser(ws, receivedMessage);
         break;
       default:
          console.log(`Uknown message type ${type}`);
